pyproject.toml
[project]
name = "tradingbot"
version = "0.1.0"
description = "A smart trading bot advisor"
authors = [{ name = "Your Name", email = "your@email.com" }]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "flask",
    "flask-jwt-extended",
    "tensorflow",
    "pandas",
    "numpy",
    "scikit-learn",
    "MetaTrader5",
    "redis",
    "pyyaml",
    "celery",
    "python-binance",
    "joblib"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
